Refactor WorkoutDetails and ExerciseViewModel
Renamed `DurationInMinutes` to `Duration` in `WorkoutDetails` for clarity. Updated `Exercises` to use `ExerciseViewModel` instead of `WorkoutExerciseViewModel`.

In `ExerciseViewModel`, renamed `ExerciseId` to `Id` and `RestTimeInSeconds` to `RestTime`. Removed `TargetMuscles` and `EquipmentNeeded` properties. Retained and reordered the `Order` property.

Ensured `Exercises` collection is still ordered by `Order` before conversion to a list.